+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 - You can now define a configuration module with `use Flop` to set defaults
   instead of or in addition to the application configuration. This makes it
   easier to work with multiple Ecto repos.
-- The new function `Flop.bindings/2` returns the necessary bindings for a
+- The new function `Flop.bindings/3` returns the necessary bindings for a
   given Flop query. You can use it in case you want to optimize your queries by
   only joining tables that are actually needed.
+- Added a `count_query` option to override the count query used by
+  `Flop.run/3`, `Flop.validate_and_run/3` and `Flop.validate_and_run!/3`.
 
 ### Changed